+++ COMP3404/COMP3404-Client/Views/ChatView.xaml.cs
@@ -1,0 +1,14 @@
+namespace COMP3404_Client.Views;
+
+public partial class ChatView : ContentView
+{
+	public ChatView()
+	{
+		InitializeComponent();
+	}
+    private void SendButtonClicked(object sender, EventArgs e)
+    {
+        TTS.instance.Speak(chatInputFrame.Text);
+        chatInputFrame.Text = string.Empty;
+    }
+}
